Fix deadlock issue when boxes are deleted or updated by the maintenance routine
We found a deadlock issue when the maintenance routine under a read lock
tries to obtain a write lock to update or delete a box. Moved the box
maintenance to it's own function so we can also test it works properly.
it now returns lists for which boxes to delete and update so they can
be done when the readlock has been released.